// src/enums/liquibase-commands.enum.ts
export enum LiquibaseCommands {
  CalculateCheckSum = 'calculateCheckSum',
  ChangelogSync = 'changelogSync',
  ChangelogSyncSql = 'changelogSyncSQL',
  ChangelogSyncToTag = 'changelogSyncToTag',
  ChangelogSyncToTagSql = 'changelogSyncToTagSQL',
  Checks = 'checks',
  ClearCheckSums = 'clearCheckSums',
  DbDoc = 'dbDoc',
  Diff = 'diff',
  DiffChangeLog = 'diffChangeLog',
  DropAll = 'dropAll',
  ExecuteSql = 'executeSql',
  Flow = 'flow',
  FutureRollbackCountSql = 'futureRollbackCountSQL',
  FutureRollbackFromTagSql = 'futureRollbackFromTagSQL',
  FutureRollbackSql = 'futureRollbackSQL',
  GenerateChangeLog = 'generateChangeLog',
  Help = 'help', // custom command
  History = 'history',
  Init = 'init',
  ListLocks = 'listLocks',
  MarkNextChangeSetRan = 'markNextChangeSetRan',
  MarkNextChangeSetRanSql = 'markNextChangeSetRanSQL',
  ReleaseLocks = 'releaseLocks',
  Rollback = 'rollback',
  RollbackCount = 'rollbackCount',
  RollbackCountSql = 'rollbackCountSQL',
  RollbackOneChangeset = 'rollbackOneChangeset',
  RollbackOneChangesetSql = 'rollbackOneChangesetSql',
  RollbackOneUpdate = 'rollbackOneUpdate',
  RollbackOneUpdateSql = 'rollbackOneUpdateSql',
  RollbackSql = 'rollbackSQL',
  RollbackToDate = 'rollbackToDate',
  RollbackToDateSql = 'rollbackToDateSQL',
  Snapshot = 'snapshot',
  SnapshotReference = 'snapshotReference',
  Status = 'status',
  Tag = 'tag',
  TagExists = 'tagExists',
  UnexpectedChangeSets = 'unexpectedChangeSets',
  Update = 'update',
  UpdateCount = 'updateCount',
  UpdateCountSql = 'updateCountSQL',
  UpdateOneChangeset = 'updateOneChangeset',
  UpdateOneChangesetSql = 'updateOneChangesetSql',
  UpdateSql = 'updateSQL',
  UpdateTestingRollback = 'updateTestingRollback',
  UpdateToTag = 'updateToTag',
  UpdateToTagSql = 'updateToTagSQL',
  Validate = 'validate',
}